pull evaluate_workflow_action_filters to processors/action for a bit more clarity
